requirements.txt
streamlit
pandas
plotly
numpy
boto3
python-dotenv
